+++ oled/__init__.py
@@ -3,7 +3,7 @@
 # See LICENSE.rst for details.
 
 """
-OLED display driver for SSD1306, SSD1331 and SH1106 devices.
+OLED display driver for SSD1306, SSD1325, SSD1331 and SH1106 devices.
 """
 
-__version__ = '1.4.0'
+__version__ = '1.5.0'