+++ CHANGES.rst
@@ -5,6 +5,7 @@ ChangeLog
 | Version    | Description                                                         | Date       |
 +============+=====================================================================+============+
 | *Upcoming* | * Support for 4-bit greyscale OLED (SSD1325)                        |            |
+|            | * Landscape/portrait orientation with rotate=N parameter            |            |
 +------------+---------------------------------------------------------------------+------------+
 | **1.4.0**  | * Add savepoint/restore functionality                               | 2016/12/23 |
 |            | * Add terminal functionality                                        |            |