+++ CHANGES.rst
@@ -4,7 +4,9 @@ ChangeLog
 +------------+---------------------------------------------------------------------+------------+
 | Version    | Description                                                         | Date       |
 +============+=====================================================================+============+
-| *Upcoming* | * Performance improvements for SH1106 driver (2x frame rate!)       |            |
+| *Upcoming* | *TBD*                                                               |            |
++------------+---------------------------------------------------------------------+------------+
+| **1.5.0*** | * Performance improvements for SH1106 driver (2x frame rate!)       | 2017/01/09 |
 |            | * Support for 4-bit greyscale OLED (SSD1325)                        |            |
 |            | * Landscape/portrait orientation with rotate=N parameter            |            |
 +------------+---------------------------------------------------------------------+------------+